Import alloc::vec where needed
This requires having a heap (+ global allocator), which may not be the
case on all embedded devices, however the targeted arm cortex M does
have one (must be added via a crate). Alternatively we could change all
occurances of `Vec<T>` to a heapless vec (separate crate). This would
add some limitations back (like having to specify a maximum size
everywhere and acting on the result of any push operations) but should
remove the requirement on heap allocation. For sake of simplicity I have
not done this here.