protobuf: Remove `WatchUpdateMsg`
It was almost identical `WatchRequestMsg`, just without `Params`.
However to use the WithdrawalAuth we need the params anyways or need to
add half the information stored there to the WithdrawalAuth objects,
making the difference between the two even smaller. If we still need the
distinction between those two message types I'd propose to make them use
exactly the same fields.