remove channel reference from ChannelUpdate
Having a mutable reference to the channel caused issues when trying to
use it outside of an async function (when trying to put it into a state
machine, as that resulted in a self-referential struct.

As a side-effect this change reduces the amount of lifetimes, thus
making the code easier to read and reason about, and allows the user to
propose multiple updates, he just has to choose one of them at the end,
making all others invalid (runtime checks). Weather the last one is good
or bad I cannot say. The go-perun implementation currently requires
this, but there is nothing in the contracts or protocols that prevents
it.